Solved bug when loading huge files

+++ src/binarize_arpa.cc
@@ -55,7 +55,7 @@ namespace Arpa2Lira {
     // find size of input file
     struct stat statbuf;
     assert(fstat(file_descriptor,&statbuf) >= 0);
-    int filesize = statbuf.st_size;
+    size_t filesize = statbuf.st_size;
     // mmap the input file
     char *filemapped;
     assert((filemapped = (char*)mmap(0, filesize,
